+++ services/src/main/java/org/keycloak/broker/saml/SAMLIdentityProviderConfig.java
@@ -431,7 +431,6 @@ public class SAMLIdentityProviderConfig extends IdentityProviderModel {
     public void validate(RealmModel realm) {
         SslRequired sslRequired = realm.getSslRequired();
 
-        checkUrl(SslRequired.NONE, getEntityId(), ENTITY_ID);
         checkUrl(sslRequired, getSingleLogoutServiceUrl(), SINGLE_LOGOUT_SERVICE_URL);
         checkUrl(sslRequired, getSingleSignOnServiceUrl(), SINGLE_SIGN_ON_SERVICE_URL);
         checkUrl(sslRequired, getMetadataUrl(), METADATA_URL);