+++ themes/src/main/resources/theme/keycloak.v2/account/src/app/content/linked-accounts-page/LinkedAccountsPage.tsx
@@ -286,16 +286,46 @@ class LinkedAccountsPage extends React.Component<LinkedAccountsPageProps, Linked
                     </Flex>
                     <StackItem isFilled/>
                     {(features.manageAccountAllowed || features.manageAccountLinkAllowed) &&
-                    <StackItem isFilled>
-                        <Title headingLevel={TitleLevel.h2} size='2xl'>
-                            <Msg msgKey='unlinkedLoginProviders'/>
-                        </Title>
-                        <DataList id="unlinked-idps" aria-label='foo'>
-                            {this.makeRows(this.state.unLinkedAccounts, false)}
-                        </DataList>
-                    </StackItem>
-                
-              
+                    <React.Fragment>
+                        <StackItem isFilled>
+                            <Title headingLevel={TitleLevel.h2} size='2xl'>
+                                <Msg msgKey='unlinkedLoginProviders'/>
+                            </Title>
+                            <DataList id="unlinked-idps" aria-label='foo' isCompact>
+                                {this.makeRows(this.state.unLinkedAccounts, false)}
+                            </DataList>
+                        </StackItem>
+                        <Flex>
+                            <FlexItem>
+                                <TextInput
+                                    type="search"
+                                    value={this.state.unLinkedAccountsKeyword}
+                                    id="unlinked-search"
+                                    name="unlinked-search"
+                                    onChange={this.onUnlinkedFilterChange}
+                                    aria-label="Search unlinked"
+                                    placeholder="Filter"
+                                    onKeyDown={this.filterUnlinkedKeyPress}
+                                />
+                            </FlexItem>
+                            <FlexItem>
+                                <Button variant="control" aria-label="filter-unlinked" onClick={this.filterUnlinkedButton}>
+                                    <SearchIcon />
+                                </Button>
+                            </FlexItem>
+                            <FlexItem>
+                                <Pagination
+                                    itemCount={this.state.unLinkedAccountsHits}
+                                    widgetId="pagination-unlinked"
+                                    perPage={this.state.unLinkedAccountsPerPage}
+                                    page={this.state.unLinkedAccountsPage}
+                                    variant={PaginationVariant.bottom}
+                                    onSetPage={this.onUnlinkedSetPage}
+                                    onPerPageSelect={this.onUnlinkedPerPageSelect}
+                                />
+                            </FlexItem>
+                        </Flex>
+                    </React.Fragment>
 
                     }
                 </Stack>